refactor(core): lift diagnostics from unifi-mcp-shared to unifi-core
Re-export shim left in unifi-mcp-shared for backwards compatibility during Phase 0;
shim removed in Task 17. Shim explicitly re-exports underscore-prefixed helpers
(_redact, _safe_json, _truncate) since `import *` skips them.
